+++ src/vite-env.d.ts
@@ -1,2 +1,7 @@
 declare const __APP_VERSION__: string
-declare const __VUE_APP_BACKEND_URL__: string | undefined
+
+declare module "*.svg?component" {
+  import { FunctionalComponent, SVGAttributes } from "vue"
+  const src: FunctionalComponent<SVGAttributes>
+  export default src
+}